requirements.txt
numpy
opencv-python
picamera2
pyyaml
loguru
mysql-connector-python
